refactor: [리팩토링] 전체 당첨 금액 계산 시, 타입을 BigDecimal로 변환하여 반환

+++ src/main/kotlin/lotto/domain/LottoRanks.kt
@@ -1,14 +1,16 @@
 package lotto.domain
 
+import java.math.BigDecimal
+
 class LottoRanks(
     private val ranks: List<LottoRank>
 ) {
 
-    fun totalPrize(): Int {
+    fun totalPrize(): BigDecimal {
         return ranks.stream()
             .mapToInt { rank -> rank.prizeMoney() }
             .sum()
-            .toInt()
+            .toBigDecimal()
     }
 
 }
